pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "schemas", "services", "utils"]

[project]
name = "ia-recruitment"
version = "0.1.0"
description = "Chatbot de recrutamento e seleção com FastMCP + Agno + AgentOS"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.0",
    "alembic>=1.18.4",
    "anthropic>=0.85.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "fastmcp>=2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
